Fix a bug in Condition, whereby the Future needs to be re-submitted upon an exception. - Move Condition to the production classpath. - Import cleanup.
